I3 mypy wave 4: 265→178 (-87 errors) across 3 dispatched agents
Three agents ran in parallel:

1. RunnerMypyContinue: siglab/evaluation/runner.py 24→0 mypy errors
   - cast() for Hashable→Timestamp narrowing
   - cast() for Returning Any from Series
   - replaced ~Index[bool] with ~pd.isna(series.index)
   - cast(Any, ...) wrap for idxmax/idxmin int|str|Series

2. FeatureDslMypy: siglab/evaluation/{feature_dsl,compile}.py 20→0 mypy errors
   - cast() for Returning Any from DataFrame
   - replaced np.sign(frame).diff() with frame.apply(np.sign).diff()
   - dict[Hashable, float]→dict[str, float] via cast
   - float() wrap for floating[Any]→float narrowing

3. TuiDataMypy (cancelled mid-flight but landed partial fixes):
   - siglab/tui/screens/market.py: type-args for list/dict, raise TypeError instead
     of return item # type:ignore, **kwargs: Any annotations
   - siglab/tui/screens/evidence.py: list[dict] for graph nodes, None check before
     .append_text
   - siglab/tui/screens/strategy.py: Callable annotation, dict type narrowing
   - siglab/tui/screens/paper.py: lambda type annotation via _show_err helper
   - siglab/tui/api_client.py: cast(dict[str, Any], response.json())
   - siglab/tui/loading.py: **kwargs: Any annotation
   - siglab/tui/widgets/sparkline.py: list[dict[str, Any]] + **kwargs: Any
   - siglab/tui/widgets/status_bar.py: **kwargs: Any annotation
   - siglab/tui/data_views.py: list[float] + type-args
   - siglab/data/{evidence,feeds,sodex_feeds,sosovalue_client}.py: small fixes

Reverted: siglab/tui/widgets/base.py (agent added raise NotImplementedError
which triggered hardening_profile.py's stub detector and broke --strict profile)

Verification:
- pytest: 2713 passed / 59 skipped / 0 failed in 48s
- ruff: 0 errors (2 agent regressions fixed)
- mypy --strict: 265 → 178 errors (-87, 32.8% reduction in this wave)
- No # type:ignore or # noqa suppressions used

+++ siglab/data/sosovalue_client.py
@@ -687,7 +687,7 @@ class SoSoValueClient:
 
     def _backoff_s(self, attempt: int) -> float:
         base = min(2.0, 0.25 * (2**attempt))
-        return base + random.uniform(0.0, base * 0.25)
+        return float(base + random.uniform(0.0, base * 0.25))
 
     def _verify_config(self) -> ssl.SSLContext | bool:
         if self.verify is not None:
